Add abstract class to reduce code duplication
